Use tail-call opcode dispatch in zkEVM
